turned read_dataset into data_access_object
